chore(multi-modal): Rename frontend.py to web.py
To fix a name clash with `components/frontend`.
