+++ examples/multimodal/graphs/disagg.py
@@ -15,9 +15,9 @@
 
 from components.decode_worker import VllmDecodeWorker
 from components.encode_worker import VllmEncodeWorker
-from components.frontend import Frontend
 from components.prefill_worker import VllmPrefillWorker
 from components.processor import Processor
+from components.web import Frontend
 
 Frontend.link(Processor).link(VllmDecodeWorker).link(VllmPrefillWorker).link(
     VllmEncodeWorker
